Fix image painting when rotate !== 0
Previously the canvas was scaled before applying the rotation,
but the transform origin did not take account of the scaling,
which meant that the image was not rotated around its center,
causing it to ping off somewhere else off-canvas.  Now, the
scale factor is applied after the transform, so there's no
need to account for it.

Fixes #183.